+++ packages/@webex/internal-plugin-metrics/src/call-diagnostic/call-diagnostic-metrics.ts
@@ -16,6 +16,8 @@ import {
   extractVersionMetadata,
   isMeetingInfoServiceError,
   isBrowserMediaErrorName,
+  isNetworkError,
+  isUnauthorizedError,
 } from './call-diagnostic-metrics.util';
 import {CLIENT_NAME} from '../config';
 import {
@@ -46,6 +48,8 @@ import {
   BROWSER_MEDIA_ERROR_NAME_TO_CLIENT_ERROR_CODES_MAP,
   MEETING_INFO_LOOKUP_ERROR_CLIENT_CODE,
   CALL_DIAGNOSTIC_LOG_IDENTIFIER,
+  NETWORK_ERROR,
+  AUTHENTICATION_FAILED_CODE,
 } from './config';
 import {generateCommonErrorMetadata} from '../utils';
 
@@ -464,6 +468,26 @@ export default class CallDiagnosticMetrics extends StatelessWebexPlugin {
         clientErrorCode: MEETING_INFO_LOOKUP_ERROR_CLIENT_CODE,
         serviceErrorCode,
       });
+    }
+
+    if (isNetworkError(rawError)) {
+      const payload = this.getErrorPayloadForClientErrorCode({
+        clientErrorCode: NETWORK_ERROR,
+        serviceErrorCode,
+      });
+      payload.errorDescription = rawError.message;
+
+      return payload;
+    }
+
+    if (isUnauthorizedError(rawError)) {
+      const payload = this.getErrorPayloadForClientErrorCode({
+        clientErrorCode: AUTHENTICATION_FAILED_CODE,
+        serviceErrorCode,
+      });
+      payload.errorDescription = rawError.message;
+
+      return payload;
     }
 
     // otherwise return unkown error